Use batching controller to batch controller and compound query changes.
Also force compound query to notify of changes immediately. Doing this helps to make the compound results controller faster because there is now only the batching interval of the compound controller to wait for, we no longer need to wait for the compound query’s batching interval to fire as well.

This also fixes the problem I was having where because the compound controller and compound query both were using a batching controller with the same batching interval, the query would finish batching too late and force a double will/did change notification to fire from the compound controller delegate.