Refactored results controller change tracking.
A changes object is passed back to a `changeTracker`. The changes object can be enumerated to extract the section and row changes. This also fixes a bug with the compound results controller where diffs were being fired when not all controller had finished fetching their content. This of course caused incorrect indexPath mappings in the diff.