Fix bug where a “move” could be incorectly interpreted as a “update”
Specifically, if a row was moved to and from the same index path (e.g. if we have 2 sections, and we move the only row of section 0 to row 0 or section 1 – This results in a move from (0,0) to (0,0)), although this looks like a simple “update”, this is truly a “move”.

The incorect interpretation was caused by an erroneous conditional (ensuring that the “to” and “from” index paths were actually different) intended to be a safety check. The conditional was erroneous because we’ve already determined that the change was a “move”.