Rework batch controller based on grand central dispatch
The batch controller has been changed to use a background dispatch queue to schedule reads/writes to its batches. Updates to the batch will come in on the main thread, and the batching work is done on a background queue. Throttling has been also implemented using grand central dispatch `DispatchWorkItem` instead of using the more primitive `Timer` class. Overall by executing work on a serial queue, the behaviour is more predictable and allows more time/breathing room for all Firebase observers to finish on the main thread before closing a batch on a background queue.

Note: The `BatchingController` has also been renamed to `BatchController` in this commit